Scrapper/requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=1.0.0
langchain-google-genai
langgraph
fastapi
uvicorn
ddgs
langchain-community
langchain